+++ backend/db/schema/workflow_node_ui.sql
@@ -1,7 +1,7 @@
 CREATE TABLE workflow_node_ui (
     id INTEGER PRIMARY KEY,
-    x_position REAL NOT NULL,
-    y_position REAL NOT NULL,
+    x_position DOUBLE PRECISION NOT NULL,
+    y_position DOUBLE PRECISION NOT NULL,
     node_label TEXT,
     node_type TEXT NOT NULL,
     FOREIGN KEY (id) REFERENCES workflow_node(id) ON DELETE CASCADE
